[Cleanup] Move utils with other utilities
