+++ qek/data/__init__.py
@@ -1,10 +1,3 @@
 """
 Data manipulation utilities.
-
-Submodules:
-    dataset: Loading, saving, manipulation or processed data.
-    datatools: Splitting training data.
-    extractors: High-level API to compile graphs and process them using a Quantum Device emulator
-        or an actual QPU.
-    graphs: Low-level API for loading and compiling graphs, for use with a `backend`.
 """
