+++ qek/kernel/__init__.py
@@ -1,7 +1,7 @@
+"""
+The Quantum Evolution Kernel itself, for use in a machine-learning pipeline.
+"""
+
 from .kernel import QuantumEvolutionKernel
 
 __all__ = ["QuantumEvolutionKernel"]
-
-"""
-The machine-learning components for this library.
-"""
