+++ qek/shared/__init__.py
@@ -1,7 +1,3 @@
 """
 Shared utility code.
-
-Submodules:
-    errors: Exceptions used in this library.
-    utils: Utilities that don't really fit anywhere else.
 """
